Version 0.0.1a of the bridge
It works locally (New issue triggers an Asana Task, whose URL is posted
back on the issue as a GH Comment) -- tested with hand-set env vars.

Need to commit it and push it up so we can test it via the GH Actions
framework itself. As this is a reusable workflow, we'll also need to
add a calling workflow to trigger it.