Refactor permission check to use an allowlist, rather than inspect teams/org
When moving this project to production, we encountered a situation where
the limited scope of the a service account's token was a problem:
* Non-members of an org cannot probe org membership
* Service/non-human accounts may not be added to an org based on Mozilla's own rules

As such, this changeset pivots to use a more manual, but greatly simpler
approach where we simply have an allowlist of Github usernames for the users
whose actions are allowed to trigger the Github-Asana bridge. For an org like
Mozilla, where we work in the open, these usernames are readily deriveable
and so I'm not considering them secret, so they can be added into a GHA
workflow that uses this action.